feat: remove unused css (mostly)
This adds one of the 5 reduction steps: unused css removal. It works by
running `purgecss` against all the inline `<style>` elements found on
the page, then writing the results back. It currently does _not_ purge:
- `@font-face` or `@keyframes` rules
- empty `@media` queries
- standalone pseudo selectors (`:before`, `::-webkit-search-decoration`,
  `::backdrop`, etc.)
- `:where()` selectors

Reducing the output further will be handled by a later step that tries
to remove selectors that _do_ apply to elements to see which are
actually necessary.